Move convert func to convert.go
